fix: get current proposal info
